Split confirm component into sub components - jsdom might not be required
