+++ package.json
@@ -24,7 +24,7 @@
     "lint": "eslint './src/**/*.js' './test/**/*.js'",
     "flow": "flow check",
     "bundle-analyzer": "webpack-bundle-analyzer",
-    "test": "jest",
+    "test": "jest --config ./jest.config.js",
     "build-test": "npm run build && npm run test:ui",
     "test:ui": "babel-node test/main.js",
     "bundlesize": "bundlesize"
@@ -52,17 +52,7 @@
       "path": "./dist/onfido.crossDevice.min.js",
       "maxSize": "1.3 kB"
     }
-  ], 
-  "jest": {
-    "moduleNameMapper": {
-      "\\.(svg|woff2?|ttf|eot|jpe?g|png|gif)$": "<rootDir>/__mocks__/fileMock.js",
-      "\\.(css|less)$": "identity-obj-proxy",
-      "^react$": "<rootDir>/node_modules/preact-compat",
-      "^react-dom$": "<rootDir>/node_modules/preact-compat",
-      "^react-modal$": "<rootDir>/node_modules/react-modal-onfido",
-      "^~utils(.*)$": "<rootDir>/src/components/utils"
-    }
-  },
+  ],
   "devDependencies": {
     "@babel/cli": "^7.4.3",
     "@babel/core": "^7.4.3",
@@ -100,6 +90,7 @@
     "chalk": "^2.4.2",
     "css-loader": "^3.2.0",
     "enzyme": "^3.11.0",
+    "enzyme-adapter-preact-pure": "^2.2.0",
     "eslint": "^6.2.2",
     "eslint-plugin-flowtype": "^4.2.0",
     "eslint-plugin-react": "^7.14.3",
@@ -111,6 +102,7 @@
     "imports-loader": "^0.8.0",
     "jest": "^24.1.0",
     "jest-cli": "^24.1.0",
+    "jsdom": "^15.2.1",
     "less": "^3.9.0",
     "less-loader": "^4.0.4",
     "mini-css-extract-plugin": "^0.5.0",
