+++ src/components/utils/unit_tests/enzymeSetup.js
@@ -1,0 +1,20 @@
+// import Enzyme from 'enzyme'
+// import Adapter from 'enzyme-adapter-preact-pure';
+// import { JSDOM } from 'jsdom'
+
+// const jsdom = new JSDOM('<!doctype html><html><body></body></html>');
+// const { window } = jsdom;
+
+// function copyProps(src, target) {
+//   Object.defineProperties(target, {
+//     ...Object.getOwnPropertyDescriptors(src),
+//     ...Object.getOwnPropertyDescriptors(target),
+//   });
+// }
+
+// global.window = window;
+// global.document = window.document;
+// copyProps(window, global);
+
+
+// Enzyme.configure({ adapter: new Adapter() });
